% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscelaneous.R
\name{binormal_auc}
\alias{binormal_auc}
\title{Binormal Area Under Curve (AUC) for ROC curve}
\usage{
binormal_auc(mean0, mean1, sigma0, sigma1)
}
\arguments{
\item{mean0}{numeric value of mean for class label 0 (healthy)}

\item{mean1}{numeric value of mean for class label 1 (diseased)}

\item{sigma0}{numeric value of standard deviation for class label 0 (healthy)}

\item{sigma1}{numeric value of standard deviation for class label 1 (diseased)}
}
\value{
a numeric vector with same dimension as x with the binormal ROC values for specified values x
}
\description{
Compute the AUC for ROC curve based on  binormal model
}
