% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binormal_roc.R
\name{binormal_roc}
\alias{binormal_roc}
\title{Binormal ROC curve}
\usage{
binormal_roc(x, mean0, mean1, sigma0, sigma1)
}
\arguments{
\item{x}{numeric vector between 0 and 1}

\item{mean0}{numeric value of mean for class label 0 (healthy)}

\item{mean1}{numeric value of mean for class label 1 (diseased)}

\item{sigma0}{numeric value of standard deviation for class label 0 (healthy)}

\item{sigma1}{numeric value of standard deviation for class label 1 (diseased)}
}
\value{
a numeric vector with same dimension as x with the binormal ROC values for specified values x
}
\description{
Compute the ROC curve based on  binomal data.
}
