% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscelaneous.R
\name{binormal_roc_strat}
\alias{binormal_roc_strat}
\title{Binormal ROC curve for stratified data}
\usage{
binormal_roc_strat(
  x,
  freq_strata,
  mean0_strat,
  mean1_strat,
  sigma0_strat,
  sigma1_strat,
  by_strata = FALSE
)
}
\arguments{
\item{x}{numeric vector between 0 and 1}

\item{freq_strata}{numeric vector containing frequency of strata}

\item{mean0_strat}{numeric vector of mean for class label 0 (healthy) for all strata}

\item{mean1_strat}{numeric vector of mean for class label 1 (diseased) for all strata}

\item{sigma0_strat}{numeric vector of standard deviation for class label 0 (healthy) for all strata}

\item{sigma1_strat}{numeric vector of standard deviation for class label 1 (diseased) for all strata}

\item{by_strata}{set to TRUE if ROC curve per strata is desired}
}
\value{
a dataframe containing
\itemize{
  \item strata: if by_strata = TRUE
  \item x: values specified for x
  \item roc: values of binormal roc for each value of specified x
}
}
\description{
Compute the combined binormal ROC curve for stratified data
}
