% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_roc.R
\name{compute_auc}
\alias{compute_auc}
\title{Function to compute Area Under the Curve using trapezoid method}
\usage{
compute_auc(fpr, tpr)
}
\arguments{
\item{tpr}{}
}
\value{

}
\description{
Function to compute Area Under the Curve using trapezoid method
}
