% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sim_quant.R
\name{compute_sim_quant}
\alias{compute_sim_quant}
\title{Simulation Results}
\usage{
compute_sim_quant(sim, param_pop)
}
\arguments{
\item{sim}{simulation object}

\item{param_pop}{data frame containing finite population parameters (prob, mu0, mu1, sigma0, sigma1)}
}
\value{
a dataframe containing containing:
\itemize{
  \item fpr: False positive rate (equals to grid, if specified)
  \item tpr_th: True positive rate
  \item rel_bias_pop: cutoffs for continuous marker
  \item rel_bias_th: cutoffs for continuous marker
  \item emp_sd_roc: cutoffs for continuous marker
  \item rel_bias_pop: cutoffs for continuous marker
  \item rel_bias_pop: cutoffs for continuous marker
}
}
\description{
Function to compute relative bias, empirical standard deviation and coverage probability
referent to finite population and super population parameters
}
