% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_density_x.R
\name{estimate_density_x}
\alias{estimate_density_x}
\title{Function to estimate survey weighted density}
\usage{
estimate_density_x(x, y, weight = NULL)
}
\arguments{
\item{x}{numeric value for the point of interest}

\item{y}{numeric vector of observations from the distribution whose density is to be estimated}

\item{weight}{numeric vector of sampling weights}
}
\value{
a dataframe containing
\itemize{
  \item f_hat: Estimated density at point x specified
  \item bandwidth: Computed bandwidth
}
}
\description{
Computes the bandwidth and estimate Gaussian kernel density estimation
}
