% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mu_pi1.R
\name{estimate_mu_pi1}
\alias{estimate_mu_pi1}
\title{Estimate mu_pi1}
\usage{
estimate_mu_pi1(weight)
}
\arguments{
\item{weight}{vector of sampling weights}
}
\value{
numerical value for the mu_pi1
}
\description{
Function to estimate mu_pi1 using Hájek estimator
}
