% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_survey_roc.R
\name{estimate_survey_roc}
\alias{estimate_survey_roc}
\title{Estimate ROC curve for complex survey data}
\usage{
estimate_survey_roc(data, grid = NULL, var_weight = NULL)
}
\arguments{
\item{data}{data frame containing continuous marker X and class label D}

\item{grid}{optional set of points for roc curve}

\item{var_weight}{string variable name for sampling weight. If specified, should be included in data}
}
\value{
a dataframe containing
\itemize{
  \item fpr: False positive rate (equals to grid, if specified)
  \item tpr: True positive rate
  \item cutoff: cutoffs for continuous marker
}
}
\description{
Function to estimate ROC curve (FPR, TPR) for complex survey data
}
