% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cluster.R
\name{genenate_cluster}
\alias{genenate_cluster}
\title{Generate one cluster of binormal distributions}
\usage{
genenate_cluster(
  size_clus,
  corr_mvn = 0,
  mean0,
  mean1,
  sigma0,
  sigma1,
  p1,
  tau = 1
)
}
\arguments{
\item{size_clus}{size of cluster}

\item{corr_mvn}{numeric value for correlation in multivariate normal distribution}

\item{mean0}{numeric vector of mean for class label 0 (healthy) for all strata}

\item{mean1}{numeric vector of mean for class label 1 (diseased) for all strata}

\item{sigma0}{numeric vector of standard deviation for class label 0 (healthy) for all strata}

\item{sigma1}{numeric vector of standard deviation for class label 1 (diseased) for all strata}

\item{p1}{numeric value for proportion of diseased observations in the population}
}
\value{
Dataframe containing
\itemize{
  \item cluster_size: cluster size
  \item X: continuous marker values
  \item D: class labels (0: healthy; 1: diseased)
}
}
\description{
Generate one cluster containing healthy and diseased observations generated from normal distributions.
The correlated binary data is generated using the multivariate probit method.
The correlated continuous marker values are generated by including a random effect
}
