% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsscs.R
\name{generate_clust_size}
\alias{generate_clust_size}
\title{Generate cluster size with specified mininum size allowed}
\usage{
generate_clust_size(n_cluster, min, lambda)
}
\arguments{
\item{n_cluster}{numeric value for number of clusters to be generated}

\item{min}{minimum allowed for cluster size}

\item{lambda}{numeric value for expected cluster size for poisson distr}
}
\description{
Generate cluster size with specified mininum size allowed
}
