% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop_clust.R
\name{generate_pop_clust}
\alias{generate_pop_clust}
\title{Generate a population of clustered observations}
\usage{
generate_pop_clust(
  cluster_sizes,
  mean0,
  mean1,
  sigma0,
  sigma1,
  p1,
  corr_mvn = 0.5,
  tau = 1
)
}
\arguments{
\item{mean0}{numeric value of mean for class label 0 (healthy)}

\item{mean1}{numeric value of mean for class label 1 (diseased)}

\item{sigma0}{numeric value of standard deviation for class label 0 (healthy)}

\item{sigma1}{numeric value of standard deviation for class label 1 (diseased)}

\item{p1}{numeric value for proportion of diseased observations in the population}

\item{corr_mvn}{numeric value for correlation in multivariate normal distribution}

\item{size_clus}{size of cluster}
}
\value{
Dataframe containing
\itemize{
  \item cluster_size: cluster size
  \item X: continuous marker values
  \item D: class labels (0: healthy; 1: diseased)
}
}
\description{
Generate a population containing clusters containing healthy and diseased observations
generated from normal distributions.
The correlated binary data is generated using the multivariate probit method.
The correlated continuous marker values are generated by including a random effect
}
