% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsscs.R
\name{generate_pop_clust_strat}
\alias{generate_pop_clust_strat}
\title{Generate a stratified population of clustered observations}
\usage{
generate_pop_clust_strat(
  N,
  freq_strata,
  mean0_strat,
  mean1_strat,
  sigma0_strat,
  sigma1_strat,
  p1,
  cluster_size,
  tau
)
}
\arguments{
\item{mean0_strat}{numeric vector of mean for class label 0 (healthy) for all strata}

\item{mean1_strat}{numeric vector of mean for class label 1 (diseased) for all strata}

\item{sigma0_strat}{numeric vector of standard deviation for class label 0 (healthy) for all strata}

\item{sigma1_strat}{numeric vector of standard deviation for class label 1 (diseased) for all strata}

\item{p1}{numeric value for proportion of diseased observations in the population}

\item{cluster_size}{list of cluster sizes per strata}

\item{corr_mvn}{numeric value for correlation in multivariate normal distribution}
}
\value{
Dataframe containing
\itemize{
  \item cluster_size: cluster size
  \item X: continuous marker values
  \item D: class labels (0: healthy; 1: diseased)
}
}
\description{
Generate a population containing clusters containing healthy and diseased observations
generated from normal distributions.
The correlated binary data is generated using the multivariate probit method.
The correlated continuous marker values are generated by including a random effect
}
