% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop_strat.R
\name{generate_pop_strat}
\alias{generate_pop_strat}
\title{Generate stratified population from mixture of normal distribution}
\usage{
generate_pop_strat(
  N,
  freq_strata,
  mean0_strat,
  mean1_strat,
  sigma0_strat,
  sigma1_strat,
  p1
)
}
\arguments{
\item{N}{total size of population}

\item{freq_strata}{numeric vector containing frequency of strata}

\item{mean0_strat}{numeric vector of mean for class label 0 (healthy) for all strata}

\item{mean1_strat}{numeric vector of mean for class label 1 (diseased) for all strata}

\item{sigma0_strat}{numeric vector of standard deviation for class label 0 (healthy) for all strata}

\item{sigma1_strat}{numeric vector of standard deviation for class label 1 (diseased) for all strata}

\item{p1}{proportion of diseased observations in the population}
}
\value{
Dataframe containing
\itemize{
  \item strata: strata identifier
  \item X: continuous marker values
  \item D: class labels (0: healthy; 1: diseased)
}
}
\description{
Generate stratified population containing healthy and diseased subpopulation generated from normal distributions
}
\examples{
generate_pop_strat(N = 1000,
                   freq_strata = rep(.25, 4),
                   mean0_strat = rep(0, 4),
                   mean1_strat = 1:4,
                   sigma0_strat = rep(1, 4),
                   sigma1_strat = rep(1, 4),
                   p1 = .25)

}
