% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_roc.R
\name{plot_survey_roc}
\alias{plot_survey_roc}
\title{Funtion to compute and plot the survey roc}
\usage{
plot_survey_roc(data, var_X, var_D, var_weight = NULL, ci = FALSE)
}
\arguments{
\item{ci}{}
}
\value{

}
\description{
Funtion to compute and plot the survey roc
}
