% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsscs.R
\name{rpois_min}
\alias{rpois_min}
\title{Random generator of Poisson distribution with mininum value}
\usage{
rpois_min(min, lambda)
}
\arguments{
\item{min}{minimum allowed for cluster size}

\item{lambda}{numeric value for expected cluster size for poisson distr}
}
\description{
Random generator of Poisson distribution with mininum value
}
