% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsscs.R
\name{simulate_cum_tscs_strat2}
\alias{simulate_cum_tscs_strat2}
\title{Title}
\usage{
simulate_cum_tscs_strat2(pop, sample_size_psu, sample_size_ssu, grid = NULL)
}
\arguments{
\item{grid}{}
}
\value{

}
\description{
Title
}
