% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sample_strat.R
\name{simulate_sample_strat}
\alias{simulate_sample_strat}
\title{Simulate stratified Simple Random Samples from a population}
\usage{
simulate_sample_strat(pop, N_sample, sample_size, grid, var_weight = "weight")
}
\arguments{
\item{pop}{dataframe containing population information (strata, X, D)}

\item{N_sample}{numeric value for number of samples to be drawn}

\item{sample_size}{sample size of samples to be drawn}

\item{grid}{numeric vector of points for roc curve from 0 to 1}

\item{var_weight}{string value indicating sampling weight variable name}
}
\value{
a dataframe containing containing survey weighted point and interval estimates for ROC curve for each
sample drawn from the specified population
}
\description{
Function to generate stratified SRS from a given population and compute ROC curve for each sample drawn
}
