% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsscs.R
\name{simulate_sample_tscs_strat}
\alias{simulate_sample_tscs_strat}
\title{Title}
\usage{
simulate_sample_tscs_strat(
  pop,
  N_sample,
  sample_size_psu,
  sample_size_ssu,
  grid = NULL
)
}
\arguments{
\item{grid}{}
}
\value{

}
\description{
Title
}
