% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_strat.R
\name{simulate_strat}
\alias{simulate_strat}
\title{Simulation function for stratified SRS}
\usage{
simulate_strat(
  N_pop,
  pop_size,
  param_pop,
  prop_disease,
  N_sample,
  sample_size,
  grid = seq(0, 1, by = 0.1)
)
}
\arguments{
\item{N_pop}{numeric value for number of stratified population to be generated}

\item{pop_size}{numeric value for population size of populations generated}

\item{param_pop}{data frame containing population parameters (prob, mu0, mu1, sigma0, sigma1)}

\item{N_sample}{numeric value for number of samples to be drawn}

\item{sample_size}{numeric value for sample size of samples to be drawn}

\item{grid}{numeric vector of points for roc curve from 0 to 1}
}
\value{
a dataframe containing containing survey weighted point and interval estimates for ROC curve for each
sample drawn from each of the generated populations
}
\description{
Generates stratified populations and draw stratified SRS from each population.
Computes survey weighted ROC curve for each sample drawn form each of the genereated populations
}
