% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsscs.R
\name{simulate_tscs_strat}
\alias{simulate_tscs_strat}
\title{Simulation function for stratified Two Stage Cluster Sampling}
\usage{
simulate_tscs_strat(
  N_pop,
  N,
  param_pop,
  p1,
  cluster_size,
  tau = 1,
  N_sample,
  sample_size_psu,
  sample_size_ssu,
  grid = seq(0, 1, by = 0.1)
)
}
\arguments{
\item{N_pop}{numeric value for number of stratified population to be generated}

\item{param_pop}{data frame containing population parameters (prob, mu0, mu1, sigma0, sigma1)}

\item{N_sample}{numeric value for number of samples to be drawn from each population}

\item{sample_size_psu}{numeric value for number of psu selected per strata}

\item{sample_size_ssu}{numeric value for number of ssu selected per selected psu}

\item{grid}{numeric vector of points for roc curve from 0 to 1}

\item{N_psu}{numeric vector of number os psu's (clusters) per strata}

\item{cluster_size_strat}{list of cluster sizes per strata}

\item{prop_disease}{numeric value for the proportion of diseased in population}
}
\value{
a dataframe containing containing survey weighted point and interval estimates for ROC curve for each
sample drawn from each of the generated populations
}
\description{
Generates stratified populations and draw stratified SRS from each population.
Computes survey weighted ROC curve for each sample drawn form each of the genereated populations
}
