% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsscs.R
\name{generate_tscs_strat}
\alias{generate_tscs_strat}
\title{Generate a Stratified Two-Stage Cluster sampling}
\usage{
generate_tscs_strat(population, n_psu, n_ssu)
}
\arguments{
\item{population}{dataframe containing population}

\item{n_psu}{numeric value for number of psu to be drawn from each strata}

\item{n_ssu}{numeric value for number of ssu to be drawn from each psu}
}
\description{
Generate a Stratified Two-Stage Cluster sampling
}
